requirements.txt
SQLAlchemy
bcrypt
PyJWT
pymysql